requirements.txt
fastapi
uvicorn[standard]
langchain
faiss-cpu
pdfplumber
openai
python-dotenv
requests
beautifulsoup4
lxml

